Add files via upload
Exploratory data analysis regarding the impact of coronavirus (COVID-19) pandemic both worldwide and in the United States, performed using MS SQL Server. Several queries are created to generate visualizations in Tableau.

The data used is open source and was obtained from ourworldindata.org/covid-deaths. The data collected is current as of 16 August 2022.