Add files via upload
This highlights the data cleaning I performed for my analysis to discover what relationship exists between median income and housing affordability (or lack thereof) within the Commonwealth of Virginia, as well as to see what impact race or ethnicity may have on this relationship. Further investigations and visualizations will be generated and linked as completed.